chore(flip): drop obsolete canonical-hash floor, move empty-pattern factory to core, refresh stale comments
