+++ src/features/sequencer/lib/helpers.ts
@@ -1,61 +1,5 @@
-import { STEP_COUNT } from "@/core/audio/engine/constants";
-import { Pattern, Voice } from "@/core/audio/engine/pattern-types";
+import { createEmptyPattern } from "@/core/audio/engine/pattern-types";
 import { clamp, quantize } from "@/shared/lib/utils";
-
-/**
- * Creates an empty pattern with all triggers off, velocities at 1.0, and no accents.
- * Returns the complete Pattern structure with 8 voices and variation metadata.
- */
-function createEmptyPattern(): Pattern {
-  const voices: Voice[] = [];
-
-  for (let instrumentIndex = 0; instrumentIndex < 8; instrumentIndex++) {
-    const voice: Voice = {
-      instrumentIndex,
-      variations: [
-        {
-          triggers: Array.from({ length: STEP_COUNT }, () => false),
-          velocities: Array.from({ length: STEP_COUNT }, () => 1),
-          timingNudge: 0,
-          ratchets: Array.from({ length: STEP_COUNT }, () => false),
-          flams: Array.from({ length: STEP_COUNT }, () => false),
-        },
-        {
-          triggers: Array.from({ length: STEP_COUNT }, () => false),
-          velocities: Array.from({ length: STEP_COUNT }, () => 1),
-          timingNudge: 0,
-          ratchets: Array.from({ length: STEP_COUNT }, () => false),
-          flams: Array.from({ length: STEP_COUNT }, () => false),
-        },
-        {
-          triggers: Array.from({ length: STEP_COUNT }, () => false),
-          velocities: Array.from({ length: STEP_COUNT }, () => 1),
-          timingNudge: 0,
-          ratchets: Array.from({ length: STEP_COUNT }, () => false),
-          flams: Array.from({ length: STEP_COUNT }, () => false),
-        },
-        {
-          triggers: Array.from({ length: STEP_COUNT }, () => false),
-          velocities: Array.from({ length: STEP_COUNT }, () => 1),
-          timingNudge: 0,
-          ratchets: Array.from({ length: STEP_COUNT }, () => false),
-          flams: Array.from({ length: STEP_COUNT }, () => false),
-        },
-      ],
-    };
-    voices.push(voice);
-  }
-
-  return {
-    voices,
-    variationMetadata: [
-      { accent: Array.from({ length: STEP_COUNT }, () => false) },
-      { accent: Array.from({ length: STEP_COUNT }, () => false) },
-      { accent: Array.from({ length: STEP_COUNT }, () => false) },
-      { accent: Array.from({ length: STEP_COUNT }, () => false) },
-    ],
-  };
-}
 
 /**
  * Clamps velocity to 0-1 range with 2 decimal precision for Tone.js.
